+++ CHANGELOG.md
@@ -7,13 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - coverage report pre-commit hook to check test coverage of the codebase.
+- `limberframework.database.connections.ServerConnection` class to connect to databases located on a server.
+- mysql+mysqldb driver for connecting to a MySQL database.
 
 ### Changed
 - pre-commit to only pass files to pydocstyle from the _limberframework_ folder.
 - pytest pre-commit hook to use coverage so that a coverage report is generated.
+- pgsql driver to postgresql for use with the new ServerConnection class.
 
 ### Fixed
 - missing package docstrings, i.e. in the \_\_init\_\_.py files, identified by pydocstyle.
+
+### Removed
+- `limberframework.database.connections.PostgresConnection` class with PostgreSQL connections now using the `limberframework.database.connections.ServerConnection` and postgresql driver.
 
 ## [0.2.0] - 2020-12-27
 ### Fixed